+++ package.json
@@ -11,6 +11,7 @@
     "electron-rebuild": "^1.1.5"
   },
   "dependencies": {
+    "angular": "^1.5.8",
     "bootstrap": "^3.3.7",
     "chart.js": "^2.2.0",
     "jquery": "^3.1.0",
